requirements.txt
streamlit==1.37.0
sentence-transformers==3.0.1
faiss-cpu>=1.9.0.post1
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.4
regex==2024.7.24
textstat==0.7.4
openai==1.40.6
python-dotenv==1.0.1
